class Location was divided into two other classes
